[Fix] Apt 관련 packge 명 변경
