--- happyhouse/src/main/java/com/jsg/house/Apt/controller/RestAptController.java
@@ -1,5 +1,0 @@
-package com.jsg.house.Apt.controller;
-
-public class RestAptController {
-
-}
