--- happyhouse/src/main/java/com/jsg/house/Apt/model/mapper/DongCodeMapper.java
@@ -1,5 +1,0 @@
-package com.jsg.house.Apt.model.mapper;
-
-public class DongCodeMapper {
-
-}
