--- happyhouse/src/main/java/com/jsg/house/Apt/model/service/DongCodeService.java
@@ -1,5 +1,0 @@
-package com.jsg.house.Apt.model.service;
-
-public interface DongCodeService {
-
-}
