+++ happyhouse/src/main/java/com/jsg/house/apt/controller/RestAptController.java
@@ -1,0 +1,5 @@
+package com.jsg.house.apt.controller;
+
+public class RestAptController {
+
+}
