+++ happyhouse/src/main/java/com/jsg/house/apt/model/mapper/AptDealMapper.java
@@ -1,0 +1,5 @@
+package com.jsg.house.apt.model.mapper;
+
+public class AptDealMapper {
+
+}
