+++ happyhouse/src/main/java/com/jsg/house/apt/model/mapper/DongCodeMapper.java
@@ -1,0 +1,5 @@
+package com.jsg.house.apt.model.mapper;
+
+public class DongCodeMapper {
+
+}
