+++ happyhouse/src/main/java/com/jsg/house/apt/model/service/DongCodeService.java
@@ -1,0 +1,5 @@
+package com.jsg.house.apt.model.service;
+
+public interface DongCodeService {
+
+}
